Take ownership of object fields
These likely exist in some stack somewher, and we need them to live as long
as the object validator.

Also, by default, the builder functions now do "catch unreachable". Failure
is unlikely, so this creates a nicer API. Every function now has a tryXYZ
alternative which returns the error in cases where this is desired.